+++ src/http/http.server.ts
@@ -1,23 +1,29 @@
 import { createServer, ServerResponse } from 'http';
 import { decodeWithCodec, encodeWithCodec } from '../codec/codecs.js';
 import { processRequest } from '../rpc/server.js';
-import { validateRequest } from '../rpc/shared/request.js';
 
 /** A HTTP server configuration. */
 export interface HttpServerConfig {
+  /**
+   * Instance to use for codec & procedure handler resolution.
+   *
+   * @default undefined
+   */
+  instanceID?: string;
+
+  /**
+   * API path prefix for requests.
+   *
+   * @default '/astrobase/'
+   */
+  prefix?: string;
+
   /**
    * Port number to use.
    *
    * @default 3000
    */
   port?: number;
-
-  /**
-   * API path for requests.
-   *
-   * @default '/astrobase'
-   */
-  apiPath?: string;
 }
 
 /**
@@ -26,12 +32,14 @@ export interface HttpServerConfig {
  * @param config {@linkcode HttpServerConfig}
  */
 export function serve(config?: HttpServerConfig) {
-  const apiPath = config?.apiPath ?? '/astrobase';
+  const prefix = config?.prefix ?? '/astrobase/';
   const port = config?.port ?? 3000;
 
   return createServer((req, res) => {
+    const procedure = req.url!.split(prefix)[1];
     if (
-      validate(res, req.url === apiPath, 404) &&
+      validate(res, !!req.url?.startsWith(prefix), 404) &&
+      validate(res, !!procedure, 404) &&
       validate(res, req.method === 'POST', 405) &&
       validate(res, !!req.headers['content-type'], 415)
     ) {
@@ -39,12 +47,20 @@ export function serve(config?: HttpServerConfig) {
       req.on('data', (chunk: string) => (data += chunk));
       req.on('end', async () => {
         try {
-          const reqMsg = validateRequest(
-            await decodeWithCodec(new TextEncoder().encode(data), req.headers['content-type']!),
+          const payload = await decodeWithCodec(
+            new TextEncoder().encode(data),
+            req.headers['content-type']!,
+            config?.instanceID,
           );
 
           try {
-            const resMsg = await encodeWithCodec(await processRequest(reqMsg), 'application/json');
+            const result = await processRequest({
+              instanceID: config?.instanceID,
+              jobID: 0,
+              payload,
+              procedure,
+            });
+            const resMsg = await encodeWithCodec(result, 'application/json', config?.instanceID);
             res.writeHead(200, { 'content-type': 'application/json' }).end(resMsg);
           } catch (e) {
             // TODO: appropriate response for unsupported procedures, or procedure threw an error